requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
dataclasses-json==0.6.7
exceptiongroup==1.2.1
frozenlist==1.4.1
gigachain==0.2.0
gigachain-core==0.2.0.1
gigachain-text-splitters==0.2.0
gigachat==0.1.30
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
install==1.3.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langsmith==0.1.82
lxml==5.2.2
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.5
packaging==23.2
pydantic==2.7.4
pydantic_core==2.18.4
pyTelegramBotAPI==4.20.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.31
tenacity==8.4.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.9.4
